refactor: move utilities out of the entry_points module
* Move utils out of the entry points module and move the rose-stem entry
  point in.
* Move fileinstall into its own module.
* Sort imports.
