+++ cylc/rose/platform_utils.py
@@ -17,9 +17,9 @@
 # -----------------------------------------------------------------------------
 """Interfaces for Cylc Platforms for use by rose apps.
 """
-import subprocess
 from optparse import Values
 import sqlite3
+import subprocess
 from time import sleep
 from typing import Any, Dict
 
@@ -29,7 +29,7 @@ from cylc.flow.pathutil import get_workflow_run_pub_db_path
 from cylc.flow.platforms import (
     HOST_REC_COMMAND,
     get_platform,
-    is_platform_definition_subshell
+    is_platform_definition_subshell,
 )
 from cylc.flow.rundb import CylcWorkflowDAO
