+++ setup.cfg
@@ -89,4 +89,4 @@ cylc.pre_configure =
 cylc.post_install =
     rose_opts = cylc.rose.entry_points:post_install
 rose.commands =
-    stem = cylc.rose.stem:main
+    stem = cylc.rose.entry_points:rose_stem
