+++ tests/functional/test_ROSE_ORIG_HOST.py
@@ -59,16 +59,14 @@ to investigate failing tests.
 
 """
 
-import pytest
+from pathlib import Path
 import re
 import shutil
-
-from pathlib import Path
 from uuid import uuid4
 
 from cylc.flow.hostuserutil import get_host
 from cylc.flow.pathutil import get_workflow_run_dir
-
+import pytest
 
 HOST = get_host()
