+++ tests/functional/test_reinstall.py
@@ -27,19 +27,19 @@ At each step it checks the contents of
 - ~/cylc-run/temporary-id/opt/rose-suite-cylc-install.conf
 """
 
-import pytest
-import shutil
-
 from itertools import product
 from pathlib import Path
+import shutil
 from uuid import uuid4
 
 from cylc.flow.hostuserutil import get_host
-from cylc.flow.pathutil import get_workflow_run_dir
-from cylc.rose.utilities import (
-    ROSE_ORIG_HOST_INSTALLED_OVERRIDE_STRING as ROHIOS)
 from cylc.flow.install import reinstall_workflow
+from cylc.flow.pathutil import get_workflow_run_dir
+import pytest
 
+from cylc.rose.utilities import (
+    ROSE_ORIG_HOST_INSTALLED_OVERRIDE_STRING as ROHIOS,
+)
 
 HOST = get_host()
